+++ example/lib/example/ex_zo_monochrome.dart
@@ -1,0 +1,53 @@
+import 'package:flutter/material.dart';
+
+import 'package:zo_animated_border/zo_animated_border.dart';
+
+class ExZoMonochromeBorder extends StatelessWidget {
+  const ExZoMonochromeBorder({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: <Widget>[
+        Padding(
+          padding: EdgeInsets.symmetric(horizontal: 20),
+          child: Container(
+              alignment: Alignment.centerLeft,
+              child: Text("Snake Border",
+                  style: TextStyle(
+                      color: Colors.white, fontWeight: FontWeight.bold))),
+        ),
+        SizedBox(
+          height: 20,
+        ),
+        Row(
+          children: [
+            SizedBox(
+              width: 20,
+            ),
+            ZoMonoCromeBorder(
+              trackBorderColor: Colors.white,
+              cornerRadius: 50.0,
+              borderStyle: ZoMonoCromeBorderStyle.mirror,
+              borderWidth: 5.5,
+              child: Container(
+                width: 100,
+                height: 100,
+                alignment: Alignment.center,
+                decoration: BoxDecoration(
+                  color: Colors.red,
+                  shape: BoxShape.circle,
+                ),
+                child: Text(
+                  "Mirror",
+                  style: TextStyle(color: Colors.white),
+                ),
+              ),
+            )
+          ],
+        ),
+      ],
+    );
+  }
+}