+++ example/lib/example/ex_zo_pulsating_border.dart
@@ -1,0 +1,52 @@
+import 'package:flutter/material.dart';
+
+import 'package:zo_animated_border/zo_animated_border.dart';
+
+class ExZoPulsatingBorder extends StatelessWidget {
+  const ExZoPulsatingBorder({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: <Widget>[
+        Padding(
+          padding: EdgeInsets.symmetric(horizontal: 20),
+          child: Container(
+              alignment: Alignment.centerLeft,
+              child: Text("Pulsating Border",
+                  style: TextStyle(
+                      color: Colors.white, fontWeight: FontWeight.bold))),
+        ),
+        SizedBox(
+          height: 20,
+        ),
+        Row(
+          children: [
+            SizedBox(
+              width: 20,
+            ),
+            ZoPulsatingBorder(
+              type: ZoPulsatingBorderType.pulse,
+              borderRadius: BorderRadius.circular(100),
+              pulseColor: Colors.blue,
+              child: Container(
+                alignment: Alignment.center,
+                width: 100,
+                height: 100,
+                decoration: BoxDecoration(
+                  color: Colors.red,
+                  borderRadius: BorderRadius.circular(50),
+                ),
+                child: Text(
+                  "Radar Pulse",
+                  style: TextStyle(color: Colors.white),
+                ),
+              ),
+            )
+          ],
+        ),
+      ],
+    );
+  }
+}